tagging pre1
git-svn-id: https://gnunet.org/svn/GNUnet@541 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/include/ecrs_core.h
@@ -162,7 +162,7 @@ typedef struct {
  *  match the query
  */
 int fileBlockEncode(const DBlock * data,
-		    unsigned short len,
+		    unsigned int len,
 		    const HashCode512 * query,
 		    Datastore_Value ** value);
 
@@ -171,7 +171,7 @@ int fileBlockEncode(const DBlock * data,
  * a certain block of data.
  */
 void fileBlockGetQuery(const DBlock * data,
-		       unsigned short len,
+		       unsigned int len,
 		       HashCode512 * query);
 
 
@@ -180,13 +180,13 @@ void fileBlockGetQuery(const DBlock * data,
  * a certain block of data.
  */
 void fileBlockGetKey(const DBlock * data,
-		     unsigned short len,
+		     unsigned int len,
 		     HashCode512 * key);
 
 /**
  * What is the type of the given block of data?
  */
-unsigned int getTypeOfBlock(unsigned short size,
+unsigned int getTypeOfBlock(unsigned int size,
 			    const DBlock * data);
 
 /**
@@ -200,7 +200,7 @@ unsigned int getTypeOfBlock(unsigned short size,
  * @return SYSERR if the content is invalid or
  *   the content type is not known
  */
-int getQueryFor(unsigned short size,
+int getQueryFor(unsigned int size,
 		const DBlock * data,
 		HashCode512 * query);
 
@@ -218,7 +218,7 @@ int getQueryFor(unsigned short size,
  *         query type
  */
 int isDatumApplicable(unsigned int type,
-		      unsigned short size,
+		      unsigned int size,
 		      const DBlock * data,
 		      unsigned int keyCount,
 		      const HashCode512 * keys);
